Feat/gravabits

+++ utils/codec.h
@@ -26,7 +26,7 @@
     // Estrutura que representa um par do Run-length Encoding
     typedef struct {
         int zeros; // Número de zeros antes do coeficiente não-zero
-        float valor;    // Valor do coeficiente não-zero (ou 0.0 para representar EOB)
+        int valor;    // Valor do coeficiente não-zero (ou 0 para representar EOB)
     } PAR_RLE;
 
     typedef struct {
@@ -45,8 +45,8 @@
     void extract_block_chroma420(PIXELYCBCR *image, float block[8][8], int start_x, int start_y, int width, int height, char channel);
     void reconstructBlock8x8_Y(PIXELYCBCR *dst, float block[8][8], int start_x, int start_y, int width, int height);
     void reconstructBlock8x8_CbCr420(PIXELYCBCR *dst, float block[8][8], int start_x, int start_y, int width, int height, char channel);
-    void quantizeMacroblocks(MACROBLOCO *mb_array, int macroblock_count, float quality);
-    void dequantizeMacroblocks(MACROBLOCO *mb_array, int macroblock_count, float quality);
+    void quantizeMacroblocks(MACROBLOCO *mb_array, int macroblock_count, int quality);
+    void dequantizeMacroblocks(MACROBLOCO *mb_array, int macroblock_count, int quality);
     void vectorize_macroblocks(MACROBLOCO *macroblocks, MACROBLOCO_VETORIZADO *vectorized_macroblocks, int macroblock_count);
     void devectorize_macroblocks(MACROBLOCO_VETORIZADO *vectorized_macroblocks, MACROBLOCO *macroblocks, int macroblock_count);
     void rle_encode_macroblocks(MACROBLOCO_RLE_DIFERENCIAL *rle_macroblocks, MACROBLOCO_VETORIZADO *vectorized_macroblocks, int macroblock_count);
